gcf_video_summary/gcf1_detect_shots/requirements.txt
# https://pypi.org/project/google-cloud-storage/
google-cloud-storage==1.28.1

# https://pypi.org/project/google-cloud-videointelligence/
google-cloud-videointelligence==1.14.0
